Agrega confirmaciones Doom y Anaconda

+++ 3_user_apps.sh
@@ -40,27 +40,33 @@ flatpak --user install flathub com.github.neithern.g4music -y
 flatpak --user install flathub com.axosoft.GitKraken -y
 
 # Doom Emacs
-if [ -d ~/.emacs.d ]; then
-    rm -Rf ~/.emacs.d
+read -rp "Instalar Doom Emacs? (S/N): " DOOM
+if [[ $DOOM =~ ^[Ss]$ ]]; then
+    if [ -d ~/.emacs.d ]; then
+        rm -Rf ~/.emacs.d
+    fi
+    go install golang.org/x/tools/gopls@latest
+    go install github.com/fatih/gomodifytags@latest
+    go install github.com/cweill/gotests/...@latest
+    go install github.com/x-motemen/gore/cmd/gore@latest
+    go install golang.org/x/tools/cmd/guru@latest
+    pip install nose
+    git clone --depth 1 https://github.com/hlissner/doom-emacs ~/.emacs.d
+    ~/.emacs.d/bin/doom install
+    sleep 5
+    rm -rf ~/.doom.d
 fi
-go install golang.org/x/tools/gopls@latest
-go install github.com/fatih/gomodifytags@latest
-go install github.com/cweill/gotests/...@latest
-go install github.com/x-motemen/gore/cmd/gore@latest
-go install golang.org/x/tools/cmd/guru@latest
-pip install nose
-git clone --depth 1 https://github.com/hlissner/doom-emacs ~/.emacs.d
-~/.emacs.d/bin/doom install
-sleep 5
-rm -rf ~/.doom.d
 
 # Tmux Plugin Manager
 git clone https://github.com/tmux-plugins/tpm ~/.config/tmux/plugins/tpm
 
 # Anaconda
-# wget https://repo.anaconda.com/archive/Anaconda3-2023.03-Linux-x86_64.sh
-# chmod +x Anaconda3-2023.03-Linux-x86_64.sh
-# ./Anaconda3-2023.03-Linux-x86_64.sh
+read -rp "Instalar Anaconda3? (S/N): " ANA
+if [[ $ANA =~ ^[Ss]$ ]]; then
+    wget https://repo.anaconda.com/archive/Anaconda3-2023.03-Linux-x86_64.sh
+    chmod +x Anaconda3-2023.03-Linux-x86_64.sh
+    ./Anaconda3-2023.03-Linux-x86_64.sh
+fi
 
 # Bash
 bash -c "$(curl -fsSL https://raw.githubusercontent.com/ohmybash/oh-my-bash/master/tools/install.sh)"
